Added edge based editing
CanvasObjects contains the classes for points/lines/faces that keep the references to connected objects as well as setting the necessary event handlers.
Place, move, select points on the canvas with the mouse.
Lines can be created from the highlighted point to the mouse click location or with SHIFT to an existing point.
An added line automatically checks, if a valid face can be created.
Lines and points can be selected and deleted with DEL.
Press control while placing new points to prevent the automatic line.
Methods to convert from and to meshes are implemented.
SVG export works as it should.
Hopefully fixed the anti-clockwise ordering of points.